robo/dashboard_comercial/gerenciador_vendas/requirements.txt
Django==5.2.4
djangorestframework==3.17.1
drf-spectacular==0.29.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
python-decouple==3.8
requests==2.31.0
selenium==4.36.0
Pillow==10.1.0
